+++ server/index.js
@@ -1,8 +1,12 @@
 const path = require('path')
 const express = require('express')
 const bodyParser = require('body-parser')
+const { parse } = require('json2csv')
 const { searchContributors, searchCommittees } = require('./lib/search')
-const { getCandidateSummary } = require('./lib/queries')
+const {
+  getCandidateSummary,
+  getCandidateContributions,
+} = require('./lib/queries')
 const { getClient } = require('./db')
 const app = express()
 app.use(bodyParser.json())
@@ -18,6 +22,22 @@ const handleError = (error, res) => {
       ? `unable to process request: ${error.message}`
       : 'unable to process request'
   res.send({ error: message })
+}
+
+/**
+ * Takes an array of data and sends a CSV file
+ * The field names of the first item in the array will be the header row
+ * @param {Array<Object>} data
+ * @param {String} filename
+ * @param {import('express').Response} res
+ */
+const sendCSV = (data, filename, res) => {
+  res.setHeader('Content-type', 'application/octet-stream')
+  res.setHeader('Content-disposition', `attachment; filename=${filename}`)
+
+  const fields = Object.keys(data[0])
+  const csv = parse(data, { fields })
+  res.send(csv)
 }
 
 const api = express.Router()
@@ -92,7 +112,7 @@ api.get('/candidate/:ncsbeID/contributions', async (req, res) => {
   let client = null
   try {
     let { ncsbeID = '' } = req.params
-    const { limit = 50, offset = 0 } = req.query
+    const { limit = 50, offset = 0, toCSV = false } = req.query
     ncsbeID = decodeURIComponent(ncsbeID)
     if (!ncsbeID) {
       res.status(500)
@@ -100,50 +120,41 @@ api.get('/candidate/:ncsbeID/contributions', async (req, res) => {
         error: 'empty ncsbeID',
       })
     }
-
     client = await getClient()
-    const contributionsPromise = client.query(
-      `select count(*) over () as full_count,
-       source_contribution_id,
-       contributor_id,
-       transaction_type,
-       committee_sboe_id,
-       report_name,
-       date_occurred,
-       account_code,
-       amount,
-       form_of_payment,
-       purpose,
-       candidate_or_referendum_name,
-       declaration,
-       id,
-       name,
-       city,
-       state,
-       zip_code,
-       profession,
-       employer_name
-       from contributions
-              join contributors c on contributions.contributor_id = c.id
-      where lower(contributions.committee_sboe_id) = lower($1)
-      limit $2
-      offset $3`,
-      [ncsbeID, limit, offset]
-    )
 
-    const summaryPromise = getCandidateSummary(ncsbeID, client)
+    if (!toCSV) {
+      const contributionsPromise = getCandidateContributions({
+        ncsbeID,
+        limit,
+        offset,
+        client,
+      })
 
-    const [contributions, summary] = await Promise.all([
-      contributionsPromise,
-      summaryPromise,
-    ])
+      const summaryPromise = getCandidateSummary(ncsbeID, client)
 
-    return res.send({
-      data: contributions.rows,
-      count:
-        contributions.rows.length > 0 ? contributions.rows[0].full_count : 0,
-      summary,
-    })
+      const [contributions, summary] = await Promise.all([
+        contributionsPromise,
+        summaryPromise,
+      ])
+
+      return res.send({
+        data: contributions.rows,
+        count:
+          contributions.rows.length > 0 ? contributions.rows[0].full_count : 0,
+        summary,
+      })
+    } else {
+      const contributions = await getCandidateContributions({ ncsbeID, client })
+
+      if (contributions.rows.length < 1) {
+        return handleError(
+          new Error(`no results found for candidate with id: ${ncsbeID}`),
+          res
+        )
+      }
+
+      sendCSV(contributions.rows, `${ncsbeID}_contributions.csv`, res)
+    }
   } catch (error) {
     handleError(error, res)
   } finally {
