+++ server/lib/queries.js
@@ -25,6 +25,53 @@ const getCandidateSummary = async (ncsbeID, client) => {
   return summary.rows.length > 0 ? summary.rows[0] : {}
 }
 
+/**
+ *
+ * @param {Object} args
+ * @param {string} args.ncsbeID
+ * @param {Number|string} args.limit
+ * @param {Number|string} args.offset
+ * @param {boolean} args.asCSV
+ * @param {import('pg').PoolClient} args.client
+ * @returns {Promise<import('pg').QueryResult>}
+ */
+const getCandidateContributions = ({
+  ncsbeID,
+  limit = null,
+  offset = null,
+  client,
+}) => {
+  return client.query(
+    `select count(*) over () as full_count,
+       source_contribution_id,
+       contributor_id,
+       transaction_type,
+       committee_sboe_id,
+       report_name,
+       date_occurred,
+       account_code,
+       amount,
+       form_of_payment,
+       purpose,
+       candidate_or_referendum_name,
+       declaration,
+       id,
+       name,
+       city,
+       state,
+       zip_code,
+       profession,
+       employer_name
+       from contributions
+              join contributors c on contributions.contributor_id = c.id
+      where lower(contributions.committee_sboe_id) = lower($1)
+      limit $2
+      offset $3`,
+    [ncsbeID, limit, offset]
+  )
+}
+
 module.exports = {
   getCandidateSummary,
+  getCandidateContributions,
 }
