+++ server/package.json
@@ -17,6 +17,7 @@
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "json2csv": "^5.0.1",
     "node-pg-migrate": "^5.3.0",
     "pg": "^8.3.0",
     "pg-copy-streams": "^5.1.1"
